fix: print table summary on verbose status
