+++ packages/amplify-cli-core/src/cliViewAPI.ts
@@ -34,7 +34,7 @@ export class ViewResourceTableParams {
   getCategoryFromCLIOptions(cliOptions: object) {
     if (cliOptions) {
       return Object.keys(cliOptions)
-        .filter(key => key != 'verbose' && key !== 'yes')
+        .filter(key => key !== 'verbose' && key !== 'yes')
         .map(category => category.toLowerCase());
     } else {
       return [];
