+++ packages/amplify-cli/src/index.ts
@@ -51,50 +51,6 @@ process.on('unhandledRejection', error => {
   throw error;
 });
 
-const convertKeysToLowerCase = <T>(obj: Record<string, T>): Record<string, T> => {
-  const newObj = {};
-  Object.entries(obj).forEach(([key, value]) => { newObj[key.toLowerCase()] = value; });
-  return newObj;
-};
-
-const normalizeStatusCommandOptions = (input: Input): Input => {
-  const options = input.options ? input.options : {};
-  const allowedVerboseIndicators = [constants.VERBOSE, 'v'];
-  // Normalize 'amplify status -v' to verbose, since -v is interpreted as 'version'
-  allowedVerboseIndicators.forEach(verboseFlag => {
-    if (options.verboseFlag !== undefined) {
-      if (typeof options[verboseFlag] === 'string') {
-        const pluginName = (options[verboseFlag] as string).toLowerCase();
-        options[pluginName] = true;
-      }
-      delete options[verboseFlag];
-      options.verbose = true;
-    }
-  });
-
-  // Merge plugins and sub-commands as options (except help/verbose)
-  const returnInput = input;
-  if (returnInput.plugin) {
-    options[returnInput.plugin] = true;
-    delete returnInput.plugin;
-  }
-  if (returnInput.subCommands) {
-    const allowedSubCommands = [constants.HELP, constants.VERBOSE]; // list of sub-commands supported in Status
-    const inputSubCommands: string[] = [];
-    returnInput.subCommands.forEach(subCommand => {
-      // plugins are inferred as sub-commands when positionally supplied
-      if (!allowedSubCommands.includes(subCommand)) {
-        options[subCommand.toLowerCase()] = true;
-      } else {
-        inputSubCommands.push(subCommand);
-      }
-    });
-    returnInput.subCommands = inputSubCommands;
-  }
-  returnInput.options = convertKeysToLowerCase(options); // normalize keys to lower case
-  return input;
-};
-
 /**
  * Command line entry point
  */
@@ -107,11 +63,6 @@ export const run = async (startTime: number): Promise<void> => {
   if (input.command !== 'help') {
     // Checks for available update, defaults to a 1 day interval for notification
     notify({ defer: false, isGlobal: true });
-  }
-
-  // Normalize status command options
-  if (input.command === 'status') {
-    input = normalizeStatusCommandOptions(input);
   }
 
   // Initialize Banner messages. These messages are set on the server side
